breaking change: Move clustering code to misk-clustering
It was previously split between misk-core and misk. There are some fakes in there, but I'll leave that for another day.

This is a breaking change, but downstream repos take in the misk dep, and I've added the new module as an API dependency, so in practice it should be the same.