RealPager#nextPage returns null when there is no next page
The docstring of the `Pager#nextPage` interface reads: `Returns null when there are no more pages`. However, the `RealPager` implementation would always return a `Page` when called, potentially resulting in endless iteration if the caller was depending on a `null` result to stop calling for a next page.

This PR changes the behavior to return null when the Pager's `hasNext` is false.

However