Do not warn on unknown LD json fields
The warning here is of minimal value. It's normal to update definitions which might have newer fields unknown to older clients. It's not an error state. In cases where someone has fat fingered the definition, it's much more likely that they'll only notice that their definition is wrong when everything explodes anyway.