Stop exporting the kotlinBom as an api dependency
If there are downstream consumers of wisp that want to use
it in their buildscript, then the kotlinBom API dependency
automatically sets a minimum version for the kotlin runtime
libraries (in this case, to 1.7.0). This is generally not
desirable because wisp is meant to be a lightweight library
and shouldn't need to force consumers into a particular new
version of kotlin when an older one works just fine. This
is particularly true since wisp doesn't expose any other kotlin
libraries as API dependencies (i.e. the kotlin runtime is not
part of the public API for wisp).