Upgrade to jetty 10.0.11 (#2234)
* Upgrade to jetty 10.0.11

* Improve handling of websockets url convertion to okHttp

The Jetty team announced the end of Community Support for Jetty 9.x on Jun 1.  Users of Jetty 9.x will continue to see releases for security and critical issues until 9.x goes End of Life (EOL) which date will be announced soon. 

The upgrade to jetty 10 should be backwards compatible as most of the changes were done in misk. However, as a major release, it's better to proceed with caution, so we have marked the latest misk release as backward incompatible so service owners can carefully release the new jetty version to production.