Disable flaky test that's causing issues
The test tests jetty functionality to ensure connections are
being closed by opening several connections to the server and expecting
some of them are being closed. Finally, the test will close all the
connections. However, Jetty still thinks some of the connections are
open after that and will wait for them until the idle time is reached
or the graceful shutdown period is over.

The test is timing out before the graceful period is over, causing the
the flakiness of the test.

This feels like a Jetty issue. I will try to replicate it and send a bug report.