Roll back AWS lib version
The new version is causing issues for flink app deployments, as
auth credentials in the Cash App environment are not getting picked
up properly. This appears to be a regression in the AWS lib, and we
will be filing an issue against the upstream about it. However for
now let's roll back to the known good version that we were on
previously.