Add retries to SchemaMigrator.initialize
For cases where multiple JdbcModules are installed, they will race to
create the `schema_version` table in tests and one will fail. Wrapping
initialize with retries will fix this as one will succeed and on retry
the others will succeed.

Motivating use case is where a utility library (say feature flags) is
installed which relies on its own utility table being present in the
database and a separate JDBC module installed for the utility usage.