Add RPUSH, LPOP and RPOP commands to misk-redis client (#2676)
Add support for the `RPUSH`, `LPOP` and `RPOP` commands for Redis lists.

Also fix a somewhat subtle bug with the FakeRedis implementation of `LPUSH`. Previously, FakeRedis would push the entire ordered list of elements to the head of an existing list. This contrasts Redis' actual behaviour for LPUSH, which is to push each element one at a time to the head of the list.

Old behaviour:

```kotlin
redis.lpush("mylist", "a", "b", "c")
redis.lrange("mylist", 0, -1) // listOf("a", "b", "c")
```

New correct behaviour:

```kotlin
redis.lpush("mylist", "a", "b", "c")
redis.lrange("mylist", 0, -1) // listOf("c", "b", "a")
```

There is a chance that this change will break tests that relied on the incorrect behaviour of `lpush`. Such tests will need to be considered on an individual basis, but will just come down to fixing comparison order of lists.