Add Gradient2Limit ConcurrencyLimiter option
Added a new configuration section for the concurrency limiter. This configuration section allows setting several parameters for the concurrency limiter, as well as choosing which algorithm to us - either VegasLimit or Gradient2Limit.

Updated the message logged when requests are shed to include the number of requests rejected since the last time the message was logged.