Introduces `PeakGauge` (#2694)
A `PeakGauge` is a variant of a regular gauge that resets to an initial
value of 0 after a metric collection.

This is useful for accurately capturing maximum observed values over
time. In contrast to the histogram maximum which tracks the maximum
value in its sampling window. That sampling window typically covers
multiple metric collections.