Improve test support for the pause detector
My attempt to install the detector module universally in skim revealed a
couple of issues with other integration tests.

Misk's default test sleeper is the `FakeSleeper` which doesn't let a
sleep terminate unless some other thread manually moves the ticker the
appropriate amount. This is a strange default for misk to have, and we
should probably consider changing that, but that's beyond the scope of
what I want to do here (unclear how many folks might be relying on this
default today). This results in test failures utilizing the
`ServiceManager` as the detector service will never shut down as it is
blocked on a sleep call that will never terminate.

In practice, the only time we care to have a fake ticker or sleeper in
the pause detector is in `PauseDetectorTest`. Any integration test that
happens to install the pause detector (e.g. via some standard common
module) should not have to worry about managing the clock on its behalf.

To address these issues, this change introduces as `@ForPauseDetector`
annotation so that we have more direct control over the ticker/sleeper
used for the detector.