Add a function to snapshot the action scope
This will allow using the action scope on threads that are running
okhttp async executions. Those threads are not tied to the original
action thread in a coroutine-friendly way so we can't use the existing
coroutine APIs to propagate the action scope. Instead we need some
out-of-band mechanism, and this new function provides that ability.