Webactions: download file if Content-Disposition is attachment (#2797)
When calling a web action via the `/web-actions` dashboard, the results get embedded inline.

This PR makes it so that, if [the `Content-Disposition` header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Disposition) in the response is set to `attachment; filename="<some file name>"`, then the browser will download the response body as a file. This is the same behavior that most browsers would follow if they navigated to a page that had such a header in the response.

Note that this behavior will not be triggered if the header is just set to `attachment` with no filename; the correct behavior in that case would be to show a file download dialog, and I didn't want to get into that rat's nest.