Rework testFixtures FakeMetrics to use extention methods
By doing this, we no longer actually need FakeMetrics, nor FakeMetricsModule really. Instead we can use the real Metrics type. This doesn't mean that the testFixture FakeMetrics goes away entirely, instead it will hold the extension methods that are used to make testing easier.

As these are based around the prometheus CollectorRegistry, we will need to eventually port them to a micrometer backed registry as part of the migration. This will result in either new extension methods, or rewriting the uses of these ones into micrometer equivalents, as it tends to be more robust for testing purposes.