Stop jetty last when shutdown_sleep_ms is set to greater than 0
This commit, combined with a previous removal of a jetty shutdown
hook, means that shutdown_sleep_ms will sleep for N ms to allow
requests to drain.

The shutdown process with shutdown_sleep_ms=30000 will be:
1. Services that are upstream of jetty shutdown
2. Jetty is told to shutdown -- We sleep for 30 seconds
3. Jetty dependencies shutdown
4. Jetty itself shutsdown (for real this time)

The 30 second sleep gives us time to drain the requests that
are inflight, and keeping jetty alive means that our istio
sidecar remains open allowing outbound requests to suceed until
the end.