Better error messages for MiskConfig lists
Currently if a field is missing/unrecognised within a misk config list Misk will fail with a `NullPointerException`.

This happens because we are using the JSON `fieldName` to construct the path to the
offending node, but lists do not have a `fieldName` (instead they have an `index`).

This PR fixes this issue such, list properties are now identified with index and will have errors like:

- "'collection.0.power_level' not found"
- "could not find 'collection.0.name' of 'TestConfig' in missing_property_in_list-testing.yaml"