Add appName, deployment lambda for menu links
Internal tools often have permalinks per service and deployment
environment. This change make it possible to pass in a lambda instead of
a static string for an admin dashboard menu link's label and url.

This makes it possible to link to an example internal tool with a URL like below.

```kotlin
https://internal-tool.${deployment.name}.company.tld/service/$appName
```