Add a way to disable non error request logging
This makes changes to the request logging interceptor in order to allow disabling
logging of non-error requests completely.

Currently, the logging interceptor does not allow you to set the rate limit to 0
as 0 is used to specify no limit. I didn't want to break backwards compatility to
change it from 0 to -1 for no limit and 0 for no logging. Instead I've added a
new config boolean that can be used to set it to not log for non-error responses.

The reason that I've added this is that a rate limit of 1 is too high for high
volume APIs when there are a large number of pods handling requests. The rate
limiting is done on a per pod basis, so a rate limit of 1/second with 100 pods
will result in about 8.5million logs per day. I'd like a way to just turn these
off!