Simplify misk-redis implementation (#3198)
Simplifies the RealRedis implementation by always deferring
single operations to a pipeline. This incurs a little overhead 
since we have to create, sync, and close a pipeline for every
command, but significantly simplifies the client.

This change also removes the old timing proxy, so hopefully
that compensates for the added overhead of the pipelines.