Fix flushAll implementation for cluster mode redis
UnifiedJedis provides flushAll, but will have an error dispatching the
command to the cluster. The "proper" way to do this is to dispatch the
command to each primary node. We don't have information about which
nodes are primaries or replicas, so we try it for all of them, and
just catch and log the errors for when we fail to flush replicas.