Keep configuration cache disabled
The configuration was disabled on gradle 7.6 because the
`org.gradle.configuration-cache` gradle property was never read (the
correct property name for gradle 7 is `org.gradle.unsafe.configuration-cache`).
This retains the existing behaviour when upgrading to gradle 8, because
getting the misk build to be configuration-cache-compatible is going
to be a nontrivial effort.