Pretty-print the service graph upon start-up (#3285)
It can at times be difficult to reason about what the dependency graph
is in larger misk services. This change logs the ServiceGraph when the
ServiceManager is created during service start-up, in order to allow
service operators to better understand their constituent service 
dependencies.

By default, no graph is printed on service start-up, but this can be
opted into by passing a ServiceManagerConfig to 
MiskRealServiceModule.