Remove Config Misk-Web tab, use Metadata tab instead (#3312)
This removes the existing Misk-Web Config admin dashboard tab in favor of the new Metadata tab which is built with the new UI stack (kotlinx HTML and Hotwire). The Config link will still be in the admin dashboard for UX purposes for a while as people learn about the new Metadata tab and find the same YAML config information they're looking for in the new location.

This leaves the Database tab and Web Actions tab as remaining Misk-Web code to gradually replace in Misk for both security, velocity, functionality, and ongoing maintenance benefits.