CCPOKR-13331 | [SQS <> DDB] Implement new "bypass" module
Currently, SQS requires a LeaseManager setup for Leases, even if you do
not want to restrict job processing to specific pods. The default Lease
implementation uses DynamoDb. Adding a new bypass module that has leases
that always return true for acquire() and checkHeld() if the app is
running in the active region. This can be used with SQS to ignore the
required lease injection.

GitOrigin-RevId: 2a40ed1147d882cc892561688225f686038dfd66