feat: Add new param to Redis modules for specific config
Currently, the
[RedisModule](https://github.com/cashapp/misk/blob/8711bf7a6f68ed7b31237d37d811649d958486f2/misk-redis/src/main/kotlin/misk/redis/RedisModule.kt#L74)
in Misk uses a
[RedisConfig](https://github.com/cashapp/misk/blob/bdb3d56f8a418fa630433a6a8e374405c5a21f1e/misk-redis/src/main/kotlin/misk/redis/RedisConfig.kt#L7)
map where the Jedis client is initialized using the first item in the
map. The map is ordered by insertion which is based on the order of how
configurations are loaded. This setup can lead to unintended
consequences, particularly when multiple Redis clusters are provisioned
for a single environment. This adds a new constructor in `RedisModule`
and `RedisClusterModule` to improve config management by allowing
consumers to provide a specific config to use instead of letting the
module choose from a given configuration map.

GitOrigin-RevId: 5175feb11d51f78f77b723dee9943741bf59b8fd