Map gRPC.Internal to 400 in client metrics interceptor
This is done to align with the behavior of the
[ExceptionHandlingInterceptor](https://github.com/cashapp/misk/blob/8cffeac185c4b16bee379acfd112670656588d36/misk/src/main/kotlin/misk/web/exceptions/ExceptionHandlingInterceptor.kt#L176-L188)
which is based on the official doc for gRPC-to-HTTP status mappings
https://grpc.github.io/grpc/core/md_doc_http-grpc-status-mapping.html.
This will resolve inconsistencies between the metrics and actual HTTP
response.

Note: This approach maintains the consistencies for services using Misk
but might introduce discrepancies in metrics when interacting with
external gRPC services (e.g Armeria where the original mapping was
taken).
GitOrigin-RevId: aa2b18b2cc90183bae37b42e5ca7605b21ca6b53