Run BeforeEachCallbacks after resetting test fixtures in misk tests
This allows `BeforeEachCallback`s to be used for things like seeding
config data into the database without that data being deleted by the
JdbcTestFixture.

Also register LogLevelExtension as a junit extension, and register it
before MiskTestExtension, so that any LogLevel annotations in tests take
effect before MiskTestExtension runs and services are started.

GitOrigin-RevId: ce17361ce41b786037a25a7eeed38a99f6ee5737