Small fix to help prevent falsely reporting clock skew when the DB is under heavy load.
There is a semantic difference between how `NOW()` and `SYSDATE()` work
([docs](https://dev.mysql.com/doc/refman/8.4/en/date-and-time-functions.html#function_sysdate)),
mainly being that `NOW()` captures the time when the query begins
execution and can be manipulated by some variables (generally for
testing purposes) while `SYSDATE()` returns the time the function is
executed (i.e. closer to when it is returned to the client) and cannot
be manipulated.

GitOrigin-RevId: 298fb59bc79e5f6b3fed96017bd2cdd2f7a6b279