The `/proc/1/cgroup` check to set the `server_hostname` when running under docker is not working in all cases. We verified this by exec-ing into our running container and checking the proc file contents:
```
❯ docker exec -it our-container sh # on host
$ cat /proc/1/cgroup # in container
0::/
```

This change introduced a a `VITESS_HOST` variable similarly to the
`MYSQL_HOST` variable used for standard MySQL connections. I've left the
original fallback in this case in case anything else was relying on it.
Our use case only calls for connecting to VITESS, but I suspect other DB
types could suffer from the same issue.

GitOrigin-RevId: 83d56771222b1a8601fcd4e8cb3b3849dbebe2ce