## Description
When calling `JettyWebSocketServlet.service`, Jetty can throw when
trying to extract the query params. This is a [known
issue](https://stackoverflow.com/questions/44345128/jetty-java-lang-illegalargumentexception-not-valid-encoding),
and results in a stacktrace being sent in the response. Sending the
stacktrace by default seems like an unwanted behavior from a security
perspective, so I'd made a change that preserves the status code and
message, but removes the stacktrace to avoid leaking unwanted
information. I didn't feel it was necessary to add a flag for this to
`WebConfig` because in almost all cases I believe this to be the
behavior that users would want.

GitOrigin-RevId: 476ac8375f9717175359ac383d9d00acd33fd3d8