This PR optionally allows preventing live traffic from being accepted and restarting the container if the ZoneRulesProvider class is not properly loaded.
This is extending guava's `AbstractIdleService` rather than implementing
misk's `HealthCheck` because:
* an of `AbstractIdleService` will have the `/_liveness` endpoint return
503
* an extension of `AbstractIdleService` or implementation of
`HealthCheck` will have the `/_readiness` endpoint return 503

This service will make readiness and liveness both return 503 if the
ZoneRulesProvider class was not loaded properly. It's important for
liveness to return 503 because this indicates that the container is in
an unrecoverable state.

GitOrigin-RevId: 2ac39afdcbb3203a6f8bc871183fc6a1136521ee