Add a `withLock` method that enables leveraging advisory lock support in MySQL/Vitess and Postgres
This adds support for taking out an advisory lock in MySQL/Postgres by
adding a new `withLock` method that stands up a hibernate session and
stashes it in a thread local, and checking for that thread local prior
to opening a new session in `withSession`. This, together with the usage
of `PhysicalConnectionHandlingMode.IMMEDIATE_ACQUISITION_AND_HOLD`
enables us to take out an advisory lock on a Hibernate session, ensure
that session is holding the underlying SQL connection the entire time
we're using it, and execute one or more transactions within the bounds
of `withLock`, all on that same session. The connection handling
semantics are of import because advisory locks are bound to the SQL
connection, not to a transaction, so the same connection must be held
until the lock is released

GitOrigin-RevId: bddaf224dd08d6f0ac0232afe244c07ebbe68032