The `JedisCluster` constructor (specifically `ClusterConnectionProvider` within it) is currently hardcoded to attempt hydrating the cluster slot cache. This presents a couple problems for our current Jedis Guice modules: 1. They're doing IO during configuration, which should be side-effect-free 2. Some tests, namely injector tests but also unit tests generally, have no concrete redis to talk to, which necessitates various workarounds in those tests if they need the DI graph.
To get around this, we setup a solution similar to
`SessionFactoryService`: we bind the new Redis cluster services during
configuration, both as `ServiceModules` and as providers for
`UnifiedJedis` and `Redis`. The services then construct the
`UnifiedJedis` at service startup, not during injector construction /
module configuration.

There is a backwards incompatibility here: although there are no API
changes, the new method of providing the clients means that any
`@MiskTest(startService = false)` which needs a `Redis` or
`UnifiedJedis` will fail, as misk must now be started to inject them.

GitOrigin-RevId: 72802e0a90b2dd72400e53bace8f9bf49517f6ef