Allow binding custom StringConverters
## Description

Currently params had to be one of the supported types, or expose a
`valueOf` function. However at Faire, we want to offer better type
validation for our parameters. For example, we have `Token<T>` where `T`
could be something like `Brand` or `Product`. Each of these token types
has a unique prefix, so even if we have a `valueOf` function exposed we
lack the ability to validate the type is correct. With this change we
can inspect the type of the parameter, and validate it's the correct
type before converting the `String`.

Note that this is technically a breaking change because I've changed
`StringConverter` to an interface, and updated the public constructor of
`FeatureBindingFactory`, but they're small enough changes that I don't
think people will have any issues upgrading.

## Testing Strategy

I updated `QueryParamFeatureBindingTest` to show that a custom
`StringConverter` can be used to properly parse a string to a value
class, which otherwise would fail. I didn't update all the other binding
tests because the code path is the same, but can update them if people
feel strongly.

GitOrigin-RevId: ce50018634417ae91eef5dbba8d7e6d7c58c9b1f