use a shadow header x-grpc-timeout-propagate in METRICS/PROPAGATE_ONLY mode
Notable changes:
- In `METRICS_ONLY`, `PROPAGATE_ONLY` mode, do not set `grpc-timeout`
header on the outgoing request as this causes the grpc client to
actually timeout. This can be surprising to those that are only opting
in for a propagation mode. Instead pass on a shadow header
`x-grpc-timeout-propagate`, and the receiver can decide if they will
continue to propagate or enforce it.
- Choose the minimum of [requestDeadline, okHttpClient.callTimeout (only
if it's non-zero)] when setting deadline on the outgoing request.
Also record which was used in metrics.
- Throw a custom `DeadlineExceededException` in the outbound
interceptor, which if unhandled will bubble up as 504 or grpc-status=4
(DEADLINE_EXCEEDED).
- Metrics:
  - Add more buckets to `deadline_duration_ms` metric distribution
  - Skip metrics for `livenesscheckaction` and `readinesscheckaction`

GitOrigin-RevId: 756b1899a5a8b951de6db63e1a93894fb70785fa