misk-testing: create TestFixtureBindingValidator (#3440)
* misk-testing: create TestFixtureBindingValidator

TestFixtureBindingValidator is a test utility class that will go through guice-injected classes that inherit from the TestFixture interface and check whether the  clause is present. This is useful because the lack of such clause will cause the TestFixtures to not be reset between runs and cause unpredictable/flaky results.

Adding a test like the ExemplarTestFixtureValidationTest to your project will guarantee that no flaky tests will happen if testFixture modules are used properly.

* Update api